Moved client updates away from API, as should be
